+++ Investissement_WebClient.Core/InterfacesServices/IYahooDataService.cs
@@ -2,5 +2,5 @@ namespace Investissement_WebClient.Core.InterfacesServices;
 
 public interface IYahooDataService
 {
-    Task<Dictionary<string, double>> GetPrixActuelAsync(List<string> symboles);
+    Task<Dictionary<string, decimal>> GetPrixActuelAsync(List<string> symboles);
 }
