+++ Investissement_WebClient.Core/Modeles/Composition.cs
@@ -4,5 +4,5 @@ public class Composition
 {
     public int IdActifEnregistre { get; set; }
     
-    public double? Quantite { get; set; }
+    public decimal? Quantite { get; set; }
 }
